Time: 10 ms (84.31%), Space: 101.4 MB (27.45%) - LeetHub

+++ 2906-construct-product-matrix/2906-construct-product-matrix.java
@@ -3,21 +3,21 @@ class Solution {
         int m = grid.length;
         int n = grid[0].length;
         int[][] p = new int[m][n];
-        int mod = 12345;
+        long mod = 12345;
         
         long prefix=1;
 
         for(int i=0;i<m;i++){
             for(int j=0;j<n;j++){
                 p[i][j] = (int)prefix;
-                prefix = (int)((prefix*grid[i][j])%(long)mod);
+                prefix = (int)((prefix*grid[i][j])%mod);
             }
         }
         long suffix = 1;
         for(int i=m-1;i>=0;i--){
             for(int j=n-1;j>=0;j--){
-                p[i][j] = (int)((p[i][j]*suffix)%(long)mod);
-                suffix = (int)(suffix *grid[i][j]%(long)mod);
+                p[i][j] = (int)((p[i][j]*suffix)%mod);
+                suffix = (int)(suffix *grid[i][j]%mod);
             }
         }
         return p;
